Change sign filed type

+++ src/routes/orders.ts
@@ -383,9 +383,9 @@ OrdersRouter.post('/', requireLogin, requireWhitelist, async (ctx) => {
       OfferOrder: [
         { name: "offerer", type: "address" },
         { name: "collection", type: "address" },
-        { name: "nonce", type: "uint256" },
+        { name: "nonce", type: "uint8" },
         { name: "token", type: "address" },
-        { name: "amount", type: "uint256" },
+        { name: "amount", type: "uint8" },
         { name: 'price', type: 'uint256' },
         { name: "expiresAt", type: "uint256" },
         { name: "tokenIds", type: "uint256[]" },
